Auth structure revised. Routing revised. Login location changed
