+++ app/page.js
@@ -1,223 +1,82 @@
 'use client';
-import Image from 'next/image';
-import KickOffProject from '../components/KickOffProject';
-import PageLayout from '@/components/pageLayout';
-import { useSelector } from 'react-redux';
-import React, { useEffect, useState } from 'react';
 import Link from 'next/link';
-import { query, collection, onSnapshot, addDoc } from 'firebase/firestore';
-
-import { db } from '../firebase/firebase';
-export default function Home() {
-	const kickOffModalStatus = useSelector(
-		(state) => state.isStartProjectOpen.modalOpen
-	);
-
-	const [data, setData] = useState([]);
-	const [filteredData, setFilteredData] = useState([]);
-	const [selectedCategory, setSelectedCategory] = useState(null);
-
-	useEffect(() => {
-		const q = query(collection(db, 'app'));
-		const unsubscribe = onSnapshot(q, (querySnapshot) => {
-			let dataArr = [];
-			querySnapshot.forEach((doc) => {
-				const projectData = { ...doc.data(), id: doc.id };
-				const totalDonations = projectData.donations.reduce(
-					(total, donation) => total + parseInt(donation.donation),
-					0
-				);
-				dataArr.push({ ...projectData, totalDonations });
-			});
-			dataArr.sort((a, b) => b.totalDonations - a.totalDonations);
-
-			setData(dataArr);
-			setFilteredData(dataArr);
-		});
-		return () => unsubscribe();
-	}, []);
-
-	const filterHandle = (category) => {
-		setSelectedCategory(category);
-		const filteredProject = category
-			? data.filter((item) => item.category === category)
-			: data;
-		setFilteredData(filteredProject);
-	};
-
-	const buttons = [
-		<div key='all' className='p-2'>
-			<button
-				className={`border-2 rounded-full h-24 w-24 flex items-center justify-center ${
-					selectedCategory === null ? 'bg-blue-500' : ''
-				}`}
-				onClick={() => filterHandle(null)}
-			>
-				All
-			</button>
-		</div>,
-		<div key='animal' className='p-2'>
-			<button
-				className={`border-2 rounded-full h-24 w-24 flex items-center justify-center ${
-					selectedCategory === 'animal' ? 'bg-blue-500' : ''
-				}`}
-				onClick={() => filterHandle('animal')}
-			>
-				Animal
-			</button>
-		</div>,
-		<div key='sport' className='p-2'>
-			<button
-				className={`border-2 rounded-full h-24 w-24 flex items-center justify-center ${
-					selectedCategory === 'sport' ? 'bg-blue-500' : ''
-				}`}
-				onClick={() => filterHandle('sport')}
-			>
-				Sport
-			</button>
-		</div>,
-		<div key='education' className='p-2'>
-			<button
-				className={`border-2 rounded-full h-24 w-24 flex items-center justify-center ${
-					selectedCategory === 'education' ? 'bg-blue-500' : ''
-				}`}
-				onClick={() => filterHandle('education')}
-			>
-				Education
-			</button>
-		</div>,
-		<div key='denem1' className='p-2'>
-			<button
-				className={`border-2 rounded-full h-24 w-24 flex items-center justify-center ${
-					selectedCategory === 'denem1' ? 'bg-blue-500' : ''
-				}`}
-				onClick={() => filterHandle('denem1')}
-			>
-				Denem1
-			</button>
-		</div>,
-	];
-
+import PageLayout from '@/components/PageLayout';
+import Login from '@/components/Login';
+import { auth } from '@/firebase/firebase';
+const Welcome = () => {
+	console.log(auth?.currentUser);
 	return (
 		<>
 			<PageLayout>
-				{kickOffModalStatus && <KickOffProject />}
-				<main className='flex  flex-col   justify-center   h-full overflow-auto   md:px-12 px-6 py-24  w-full '>
-					<div className='grid grid-cols-12 gap-6 border-b-2 pb-12 '>
-						<div className='w-full flex flex-col justify-center gap-12 md:pt-0 sm:col-span-7 col-span-12  '>
-							<h1 className='text-6xl text- font-bold'>Project of the week</h1>
-							<div className='w-full'>
+				<div className=' w-full  bg-[#d4ee26] h-[100vh] lg:px-28  lg:py-56 py-28 '>
+					<div className='welcome-desktop w-full  h-full hidden lg:flex justify-end lg:justify-start  items-end lg:items-center  '>
+						<div className='lg:justify-self-end'>
+							{!auth?.currentUser && <Login></Login>}
+						</div>
+						<div className='lg:w-1/3  flex gap-8 flex-col lg:ml-36 ml-6 mr-6'>
+							<h1 className='text-8xl font-bold'>Givingly</h1>
+							<h2 className='text-3xl'>Supporting great causes made easy</h2>
+
+							<p>
+								We helped over 3,500 projects and causes. Sign in today and get
+								your idea kicked off or support others kick off their amazing
+								projects
+							</p>
+							<Link
+								className='bg-black w-full p-2 text-white rounded mt-4 hover:text-[#d4ee26] text-center '
+								href='/navigation'
+							>
+								Start today
+							</Link>
+						</div>
+					</div>
+
+					<div className='welcome-tablet  w-full hidden h-full md:flex  justify-end lg:hidden items-end   '>
+						<div className=' w-1/2 flex gap-8 flex-col lg:ml-36 ml-6 mr-6'>
+							<h1 className='text-8xl font-bold'>Givingly</h1>
+							<h2 className='text-3xl '>Supporting great causes made easy</h2>
+							<p>
+								We helped over 3,500 projects and causes. Sign in today and get
+								your idea kicked off or support others kick off their amazing
+								projects
+							</p>
+							<Link
+								className='bg-black w-full p-2 text-center text-white rounded mt-4 hover:text-[#d4ee26] '
+								href='/navigation'
+							>
+								Start today
+							</Link>
+						</div>
+
+						<div>{!auth?.currentUser && <Login></Login>}</div>
+					</div>
+
+					<div className='welcome-mobile w-full  flex md:hidden justify-center  px-12 h-full     '>
+						<div className=' w-full flex  flex-col justify-between  '>
+							<h1 className='text-6xl font-bold px-4 text-center'>Givingly</h1>
+
+							<div className='text-center flex flex-col px-4 w-full'>
 								{' '}
+								<h2 className='text-2xl'>Supporting great causes made easy</h2>
+								<p>
+									We helped over 3,500 projects and causes. Sign in today and
+									get your idea kicked off or support others kick off their
+									amazing projects
+								</p>
 								<Link
-									key={data[0]?.id}
-									className='block py-2 px-3 '
-									href={`/project/${data[0]?.id}`}
+									className='bg-black w-full p-2 text-white rounded mt-4 text-center hover:text-[#d4ee26] '
+									href='/navigation'
 								>
-									<Image
-										className='border-2 '
-										src={data[0]?.img}
-										width={400}
-										height={300}
-										alt='Picture of the author'
-									/>
+									Start today
 								</Link>
 							</div>
 						</div>
-						<div className='w-full flex flex-col gap-4 sm:col-span-5 col-span-12  justify-end  '>
-							<h3 className='font-bold lg:text-4xl md:text-3xl text-2xl '>
-								{data[0]?.title}
-							</h3>
-							<p className='text-sm'>{data[0]?.desc}</p>
-
-							<div className=' w-full flex flex-col gap-2 text-sm'>
-								<div className='h-2 w-full bg-gray-200 rounded'>
-									<div
-										style={{
-											maxWidth: '100%',
-											width: `${
-												(data[0]?.totalDonations / data[0]?.goal) * 100
-											}%`,
-										}}
-										className='h-2  bg-[#d4ee26] rounded'
-									></div>
-								</div>
-
-								<div className='grid grid-cols-12'>
-									<span className='col-span-11'>Raised</span>{' '}
-									<span className='col-span-1'>Goal:</span>
-								</div>
-
-								<div className='grid grid-cols-12'>
-									<span className='col-span-11'>
-										${data[0]?.totalDonations}
-									</span>{' '}
-									<span className='col-span-1'>${data[0]?.goal}</span>
-								</div>
-							</div>
-						</div>
+						<div>{!auth?.currentUser && <Login></Login>}</div>
 					</div>
-
-					<div className='my-6 flex flex-col justify-start items-start w-full '>
-						<h1 className='text-4xl text-left'>Categories</h1>
-						<div>
-							<ul className='flex gap-4 my-6'>{buttons}</ul>
-						</div>
-
-						<div className='flex flex-wrap w-full gap-8 '>
-							{filteredData.map((project) => {
-								return (
-									<Link
-										key={project.id}
-										className='block py-2 px-3 '
-										href={`/project/${project?.id}`}
-									>
-										<div
-											key={project.id}
-											className='flex flex-col shadow-lg gap-4 p-4 w-96'
-										>
-											<div className='relative w-96 h-64 '>
-												{' '}
-												<Image
-													className='rounded'
-													src={project?.img}
-													layout='fill'
-													alt='Picture of the author'
-												/>
-											</div>
-											<h1>{project.title}</h1>
-											<div className=' w-full flex flex-col gap-2 text-sm'>
-												<div className='h-2 w-full bg-gray-200 rounded'>
-													<div
-														style={{
-															maxWidth: '100%',
-															width: `${
-																(project.totalDonations / project.goal) * 100
-															}%`,
-														}}
-														className='h-2  bg-[#d4ee26] rounded'
-													></div>
-												</div>
-
-												<div className='grid grid-cols-12'>
-													<span className='col-span-11'>Raised</span>{' '}
-													<span className='col-span-1'>Goal</span>
-												</div>
-
-												<div className='grid grid-cols-12'>
-													<span className='col-span-11'>
-														${project.totalDonations}
-													</span>{' '}
-													<span className='col-span-1'>${project.goal}</span>
-												</div>
-											</div>
-										</div>
-									</Link>
-								);
-							})}
-						</div>
-					</div>
-				</main>
+				</div>
 			</PageLayout>
 		</>
 	);
-}
+};
+
+export default Welcome;
